nightSky/requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
black==22.1.0
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
coreschema==0.0.4
cryptography==36.0.2
Django==4.0.3
django-cors-headers==3.11.0
django-rest-knox==4.2.0
django-tinymce==3.4.0
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
jdatetime==4.1.0
Jinja2==3.1.1
jsonschema==4.4.0
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
pycodestyle==2.8.0
pycparser==2.21
pyparsing==3.0.7
pyrsistent==0.18.1
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
zeep==4.1.0
zipp==3.7.0
